Redefine some modules to make the code more intelligible
