+++ core.py
@@ -1,99 +1,15 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import collections
-
-import numpy as np
 import pandas as pd
 
 from hydrogens import get_CH, get_CH2, get_CH3, get_CH_double_bond
+import geometry
 
 
 # For debugging.
 # TODO: Remove it after implement logging feature
 DEBUG=False
-
-def make_dic_atname2genericname(filename):
-    """Make a dict of correspondance between generic H names and PDB names.
-
-    This dict will look like the following: {('C1', 'H11'): 'gamma1_1', ...}.
-    Useful for outputing OP with generic names (such as beta1, beta 2, etc.).
-    Such files can be found on the NMRlipids MATCH repository:
-    https://github.com/NMRLipids/MATCH/tree/master/scripts/orderParm_defs.
-
-    Parameters
-    ----------
-    filename : str
-        Filename containing OP definition
-        (e.g. `order_parameter_definitions_MODEL_Berger_POPC.def`).
-
-    Returns
-    -------
-    Ordered dictionnary
-        Keys are tuples of (C, H) name, values generic name (as described
-        above in this docstring). The use of an ordered dictionnary ensures
-        we get always the same order in the output OP.
-    """
-    dic = collections.OrderedDict()
-    try:
-        with open(filename, "r") as f:
-            for line in f:
-                # TODO: This line might have to be changed if the file contains more than
-                # 4 columns.
-                name, _, C, H = line.split()
-                dic[(C, H)] = name
-    except:
-        raise UserWarning("Can't read order parameter definition in "
-                          "file {}".format(filename))
-    return dic
-
-def init_dic_OP(universe_woH, dic_atname2genericname, dic_lipid):
-    """TODO Complete docstring.
-    """
-    ### To calculate the error, we need to first average over the
-    ### trajectory, then over residues.
-    ### Thus in dic_OP, we want for each key a list of lists, for example:
-    ### OrderedDict([
-    ###              (('C1', 'H11'), [[], [], ..., [], []]),
-    ###              (('C1', 'H12'), [[], ..., []]),
-    ###              ...
-    ###              ])
-    ### Thus each sublist will contain OPs for one residue.
-    ### e.g. ('C1', 'H11'), [[OP res 1 frame1, OP res1 frame2, ...],
-    ###                      [OP res 2 frame1, OP res2 frame2, ...], ...]
-    dic_OP = collections.OrderedDict()
-    # We also need the correspondance between residue number (resnum) and
-    # its index in dic_OP.
-    dic_corresp_numres_index_dic_OP = {}
-    # Create these sublists by looping over each lipid.
-    for key in dic_atname2genericname:
-        dic_OP[key] = []
-        # Get lipid name.
-        resname = dic_lipid["resname"]
-        selection = "resname {}".format(resname)
-        # Loop over each residue on which we want to calculate the OP on.
-        for i, residue in enumerate(universe_woH.select_atoms(selection).residues):
-            dic_OP[key].append([])
-            dic_corresp_numres_index_dic_OP[residue.resid] = i
-    if DEBUG:
-        print("Initial dic_OP:", dic_OP)
-        print("dic_corresp_numres_index_dic_OP:", dic_corresp_numres_index_dic_OP)
-    return dic_OP, dic_corresp_numres_index_dic_OP
-
-
-def make_dic_Cname2Hnames(dic_OP):
-    """TODO Complete Docstring.
-    """
-    dic = {}
-    for Cname, Hname in dic_OP.keys():
-        if Cname not in dic:
-            dic[Cname] = (Hname,)
-        else:
-            dic[Cname] += (Hname,)
-    if DEBUG:
-        print("dic_Cname2Hnames contains:", dic)
-    return dic
-
 
 
 ###
@@ -280,7 +196,7 @@ def build_all_Hs_calc_OP(universe_woH, dic_lipid, dic_Cname2Hnames,
                 ####
                 if dic_OP:
                     if (atom.name, Hname) in dic_OP:
-                        op = calc_OP(atom.position, H_coor)
+                        op = geometry.calc_OP(atom.position, H_coor)
                         # We should get here the index of the residue in dic_OP.
                         # For that we can use dic_corresp_numres_index_dic_OP
                         # (key: resnum in pdb, value: index residue in dic_OP).
@@ -639,7 +555,7 @@ def fast_build_all_Hs_calc_OP(universe_woH, begin, end,
                     Hname = dic_Cname2Hnames[Cname][counter4Hname]
                     # Calc and store OP for that couple C-H.
                     Cname_position = ts[Cname_ix+ix_first_atom_res]
-                    op = calc_OP(Cname_position, H_coor)
+                    op = geometry.calc_OP(Cname_position, H_coor)
                     # Old way: dic_OP[(Cname, Hname)].append(op)
                     if (Cname, Hname) in dic_OP:
                         dic_OP[(Cname, Hname)][lipid_ix].append(op)
@@ -653,40 +569,3 @@ def fast_build_all_Hs_calc_OP(universe_woH, begin, end,
     if DEBUG:
         print("Final dic_OP:", dic_OP)
         print()
-
-
-
-
-def calc_OP(C, H):
-    """Returns the Order Parameter of a CH bond (OP).
-
-    OP is calculated according to equation:
-
-    S = 1/2 * (3*cos(theta)^2 -1)
-
-    theta is the angle between CH bond and the z(vertical) axis:
-    z
-    ^  H
-    | /
-    |/
-    C
-
-    Inspired from a function written by @jmelcr.
-
-    Parameters
-    ----------
-    C : numpy 1D-array
-        Coordinates of C atom.
-    H : numpy 1D-array
-        Coordinates of H atom.
-
-    Returns
-    -------
-    float
-        The normalized vector.
-    """
-    vec = H - C
-    d2 = np.square(vec).sum()
-    cos2 = vec[2]**2/d2
-    S = 0.5*(3.0*cos2 - 1.0)
-    return S
